Move methods to abstract base class to share them with other clients
