+++ georocket-client-api/src/main/java/io/georocket/client/StoreClient.java
@@ -26,7 +26,7 @@ import io.vertx.core.streams.WriteStream;
  * @since 1.0.0
  * @author Michel Kraemer
  */
-public class StoreClient {
+public class StoreClient extends AbstractClient {
   /**
    * Actions to update meta data of existing chunks in the data store.
    */
@@ -52,49 +52,6 @@ public class StoreClient {
    */
   public StoreClient(HttpClient client) {
     this.client = client;
-  }
-  
-  /**
-   * Convenience method to URL-encode a string
-   * @param str the string
-   * @return the encoded string
-   */
-  protected String urlencode(String str) {
-    try {
-      return URLEncoder.encode(str, "UTF-8");
-    } catch (UnsupportedEncodingException e) {
-      throw new RuntimeException(e);
-    }
-  }
-  
-  /**
-   * Prepare a query. Generate a query path for the given search query and layer.
-   * @param query the search query (may be <code>null</code>)
-   * @param layer the layer to export (may be <code>null</code>)
-   * @return the query path
-   */
-  protected String prepareQuery(String query, String layer) {
-    if (layer == null || layer.isEmpty()) {
-      layer = "/";
-    } else {
-      layer = Stream.of(layer.split("/"))
-        .map(this::urlencode)
-        .collect(Collectors.joining("/"));
-    }
-
-    if (!layer.endsWith("/")) {
-      layer += "/";
-    }
-    if (!layer.startsWith("/")) {
-      layer = "/" + layer;
-    }
-    
-    String urlQuery = "";
-    if (query != null && !query.isEmpty()) {
-      urlQuery = "?search=" + urlencode(query);
-    }
-    
-    return layer + urlQuery;
   }
   
   /**
@@ -348,17 +305,6 @@ public class StoreClient {
     });
 
     return configureRequest(request);
-  }
-
-  /**
-   * Configure an HTTP request. The default implementation of this method does
-   * nothing. Sub-classes may override if they want to configure a request
-   * before it is sent.
-   * @param request request the request to configure
-   * @return same {@link HttpClientRequest} as given but with options set
-   */
-  protected HttpClientRequest configureRequest(HttpClientRequest request) {
-    return request;
   }
 
   /**
@@ -705,32 +651,6 @@ public class StoreClient {
       }
     });
     configureRequest(request).end();
-  }
-
-  /**
-   * Parses an HTTP response and calls the given handler with the
-   * parsed error message
-   * @param <T> the type of the handler's result
-   * @param response the HTTP response
-   * @param handler the handler to call
-   */
-  private static <T> void fail(HttpClientResponse response,
-      Handler<AsyncResult<T>> handler) {
-    fail(response, handler, ClientAPIException::parse);
-  }
-
-  /**
-   * Parses an HTTP response, maps it to an exception and then calls the
-   * given handler
-   * @param <T> the type of the handler's result
-   * @param response the HTTP response
-   * @param handler the handler to call
-   * @param map a function that maps the parsed error message to an exception
-   */
-  private static <T> void fail(HttpClientResponse response,
-      Handler<AsyncResult<T>> handler, Function<String, Throwable> map) {
-    response.bodyHandler(buffer ->
-      handler.handle(Future.failedFuture(map.apply(buffer.toString()))));
   }
 
   /**
